+++ src/main/java/calculatelogic/Calculation.java
@@ -2,7 +2,6 @@ package calculatelogic;
 
 import java.util.List;
 
-public interface CalculateLogic {
+public interface Calculation {
     int calculateTheWhole(List<String> formula);
-
 }
